Fixes after testing
.gitignore:
* ignore all .db because there might be more

backend/db.py:
* Fix: sql.connect is awaitable
+ Execute startup script
* Fix: commit is on dbw not db
* Change: tables are prefixed
* Change: async with execute br0nk so await

backend/server.py:
* Fix: comments API is singular user
+ /site/* maps to public/*
* _wakeup now returns on code change
* Close no matter what
* Fix: random quote after COMMENTS_API
+ file_handler for /site/
+ Fix: not_found was ironically not found

public/login/index.html:
* Absolute URLs because I don't trust slashes

+++ backend/db.py
@@ -1,6 +1,8 @@
+import os
 import time
 from secrets import token_bytes, token_hex, randbits
 from hashlib import sha256
+import asyncio
 import aiosqlite as sql
 
 DATABASE_FILENAME = 'scratchverifier.db'
@@ -9,30 +11,33 @@ USERS_API = 'https://api.scratch.mit.edu/users/{}'
 class Database:
     def __init__(self, session):
         loop = asyncio.get_event_loop()
-        self.dbw = sql.connect(DATABASE_FILENAME)
+        self.dbw = loop.run_until_complete(sql.connect(DATABASE_FILENAME))
         self.dbw.row_factory = sql.Row
         self.db = loop.run_until_complete(self.dbw.cursor())
+        with open(os.path.join(os.path.dirname(__file__), 'sql',
+                               'startup.sql')) as startup:
+            loop.run_until_complete(self.db.executescript(startup.read()))
         self.session = session
 
     async def close(self):
-        await self.db.commit()
+        await self.dbw.commit()
         await self.dbw.close()
 
     ### TABLE: clients ###
 
     async def client_matches(self, client_id, token):
-        async with self.db.execute('SELECT client_id FROM clients \
-WHERE client_id=? AND token=?', (client_id, token)):
-            if (await self.db.fetchone()):
-                return True
-            return False
+        await self.db.execute('SELECT client_id FROM scratchverifier_clients \
+WHERE client_id=? AND token=?', (client_id, token))
+        if (await self.db.fetchone()):
+            return True
+        return False
 
     ### TABLE: clients and sessions ###
 
     async def username_from_session(self, session_id):
-        async with self.db.execute('SELECT username FROM sessions \
-WHERE session_id=?', (session_id,)):
-            row = await self.db.fetchone()
+        await self.db.execute('SELECT username FROM scratchverifier_sessions \
+WHERE session_id=?', (session_id,))
+        row = await self.db.fetchone()
         if row is None:
             return None
 
@@ -45,17 +50,17 @@ WHERE session_id=?', (session_id,)):
             data = await resp.json()
         client_id = data['id']
         token = token_hex(32)
-        await self.db.execute('INSERT INTO clients (client_id, token, \
-username) VALUES (?, ?, ?)', (client_id, token, username))
+        await self.db.execute('INSERT INTO scratchverifier_clients (client_id, \
+token, username) VALUES (?, ?, ?)', (client_id, token, username))
         return {'client_id': client_id, 'token': token, 'username': username}
 
     async def get_client(self, session_id):
         username = self.username_from_session(session_id)
         if username is None:
             return None
-        async with self.db.execute('SELECT * FROM clients \
-WHERE username=?', (username,)):
-            row = await self.db.fetchone()
+        await self.db.execute('SELECT * FROM scratchverifier_clients \
+WHERE username=?', (username,))
+        row = await self.db.fetchone()
         if row is None:
             return None
         return dict(row)
@@ -64,14 +69,14 @@ WHERE username=?', (username,)):
         username = self.username_from_session(session_id)
         if username is None:
             return
-        await self.db.execute('UPDATE clients SET token=? \
+        await self.db.execute('UPDATE scratchverifier_clients SET token=? \
 WHERE username=?', (token_hex(32), username))
 
     async def del_client(self, session_id):
         username = self.username_from_session(session_id)
         if username is None:
             return
-        await self.db.execute('DELETE FROM clients \
+        await self.db.execute('DELETE FROM scratchverifier_clients \
 WHERE username=?', (username,))
 
     ### TABLE: sessions ###
@@ -79,12 +84,12 @@ WHERE username=?', (username,))
     async def new_session(self, username):
         while 1:
             session_id = randbits(64)
-            async with self.db.execute('SELECT session_id FROM sessions \
-WHERE session_id=?', (session_id,)):
-                if (await self.db.fetchone()) is None:
-                    break
-        await self.db.execute('INSERT INTO sessions (session_id, expiry, \
-username) VALUES (?, ?, ?)', (
+            await self.db.execute('SELECT session_id FROM \
+scratchverifier_sessions WHERE session_id=?', (session_id,))
+            if (await self.db.fetchone()) is None:
+                break
+        await self.db.execute('INSERT INTO scratchverifier_sessions \
+(session_id, expiry, username) VALUES (?, ?, ?)', (
             session_id,
             int(time.time),
             username
@@ -92,9 +97,12 @@ username) VALUES (?, ?, ?)', (
         return session_id
 
     async def get_expired(self, session_id):
-        async with self.db.execute('SELECT expiry FROM sessions \
-WHERE session_id=?', (session_id,)):
-            expiry = (await self.db.fetchone())[0]
+        await self.db.execute('SELECT expiry FROM scratchverifier_sessions \
+WHERE session_id=?', (session_id,))
+        expiry = await self.db.fetchone()
+        if expiry is None:
+            return True
+        expiry = expiry[0]
         if time.time() > expiry:
             await self.db.execute('DELETE FROM sessions \
 WHERE session_id=?', (session_id,))
@@ -110,15 +118,15 @@ WHERE session_id=?', (session_id,))
             + username.encode()
             + token_bytes()
         )
-        await self.db.execute('INSERT INTO usage (client_id, code, username, \
-expiry) VALUES (?, ?, ?, ?)', (client_id, code, username,
+        await self.db.execute('INSERT INTO scratchverifier_usage (client_id, \
+code, username, expiry) VALUES (?, ?, ?, ?)', (client_id, code, username,
                                int(time.time() + 1800)))
         return code
 
     async def get_code(self, client_id, username):
-        async with self.db.execute('SELECT code, expiry FROM usage \
-WHERE client_id=? AND username=?', (client_id, username)):
-            row = await self.db.fetchone()
+        await self.db.execute('SELECT code, expiry FROM scratchverifier_usage \
+WHERE client_id=? AND username=?', (client_id, username))
+        row = await self.db.fetchone()
         if row is None:
             return None
         if time.time() > row['expiry']:
@@ -127,5 +135,5 @@ WHERE client_id=? AND username=?', (client_id, username)):
         return row['code']
 
     async def end_verification(self, client_id, username):
-        await self.db.execute('DELETE FROM usage WHERE client_id=? \
-AND username=?', (client_id, username))
+        await self.db.execute('DELETE FROM scratchverifier_usage WHERE \
+client_id=? AND username=?', (client_id, username))
